requirements.txt
# Dark AI - Python Dependencies
# Install with: pip install -r requirements.txt

# Core Speech & Text
pyttsx3==2.90
SpeechRecognition==3.10.0

# Web & Network
requests==2.31.0
beautifulsoup4==4.12.2
flask==3.0.0
flask-cors==4.0.0

# Automation
pyautogui==0.9.54
pywhatkit==5.4
pynput==1.7.6

# Information & Search
wikipedia==1.4.0
wolframalpha==5.0.0

# System & Notifications
plyer==2.1.0
pygame==2.5.2

# Network Testing
speedtest-cli==2.1.3

# Image Processing (for intro)
Pillow==10.1.0

# Optional: Environment variables
python-dotenv==1.0.0

# AI Chatbot
google-generativeai==0.3.1
